+++ templates/attraction.html
@@ -7,7 +7,6 @@
         <link rel="stylesheet" href="/static/attraction.css" type="text/css">
     </head>
     <body>
-        <script src="/static/attraction.js"></script>
         <!-- dialog -->
         <div>
             <!-- 登入 -->
@@ -37,7 +36,7 @@
         <!-- nav -->
         <nav id="nav">
             <div id="nav_2">
-                <a id="homepage">台北一日遊</a>
+                <a id="homepage" onclick="back_to_home()">台北一日遊</a>
                 <a id="check" onclick="show_booking()">預定行程</a>
                 <a class="login_logout" id="login" onclick="show_dialog_login()">登入/註冊</a>
                 <a class="login_logout" id="logout" onclick="sign_out()">登出系統</a>
@@ -85,5 +84,6 @@
         <footer id="footer">
             <div id="copyright">COPYRIGHT © 2021 台北一日遊</div>
         </footer>
+        <script src="/static/attraction.js"></script>
     </body>
 </html>
