+++ templates/booking.html
@@ -36,12 +36,15 @@
         <!-- nav -->
         <nav id="nav">
             <div id="nav_2">
-                <a id="homepage">台北一日遊</a>
+                <a id="homepage" onclick="back_to_home()">台北一日遊</a>
                 <a id="check" onclick="show_booking()">預定行程</a>
                 <a class="login_logout" id="login" onclick="show_dialog_login()">登入/註冊</a>
                 <a class="login_logout" id="logout" onclick="sign_out()">登出系統</a>
             </div> 
         </nav>
+        <div class="loading_box">
+            <div class="loading"></div>
+        </div>
         <main>
             <div id="booking">
                 <div class="title">您好，<span id="username"></span>，待預訂的行程如下：</div>
