+++ templates/index.html
@@ -7,7 +7,6 @@
         <link rel="stylesheet" href="/static/index.css" type="text/css">
     </head>
     <body>
-        <script src="/static/index.js"></script>
         <!-- dialog -->
         <div>
             <!-- 登入 -->
@@ -36,7 +35,7 @@
         </div>
         <!-- nav -->
         <nav>
-           <a id="homepage">台北一日遊</a>
+           <a id="homepage" onclick="back_to_home()">台北一日遊</a>
            <a id="check" onclick="show_booking();">預定行程</a>
            <a class="login_logout" id="login" onclick="show_dialog_login();">登入/註冊</a>
            <a class="login_logout" id="logout" onclick="sign_out();">登出系統</a>
@@ -68,4 +67,5 @@
             <div id="copyright">COPYRIGHT © 2021 台北一日遊</div>
         </footer>
     </body>
+    <script src="/static/index.js"></script>
 </html>
